Add comprehensive tests for canvas manager and design layer functionality
- Implement test-canvas-fixes.html to validate canvas scaling, background color setting, and layer management.
- Create test_design_lookup_types.php to verify LayerController's handling of UUID and numeric ID lookups.
- Develop test_design_search.php to check accessibility and response of design search endpoints.
- Add test_design_show_with_layers.php to ensure design show endpoint returns complete design data with layers.
- Introduce test_layer_persistence.php to comprehensively test layer persistence functionality, including creation, updates, and verification of layers.

+++ backend/src/DTO/ValueObject/TextLayerProperties.php
@@ -30,15 +30,15 @@ final readonly class TextLayerProperties extends LayerProperties
 
         /**
          * Font size in pixels (1-500px)
-         * @var int $fontSize
+         * @var float $fontSize
          */
-        #[Assert\Type(type: 'integer', message: 'Font size must be an integer')]
+        #[Assert\Type(type: 'float', message: 'Font size must be an integer')]
         #[Assert\Range(
             min: 1,
             max: 500,
             notInRangeMessage: 'Font size must be between 1 and 500 pixels'
         )]
-        public int $fontSize = 16,
+        public float $fontSize = 16,
 
         /**
          * Font weight (normal, bold, or numeric values 100-900)
